requirements.txt
tensorflow
numpy
opencv-python
pillow
python-telegram-bot==13.15
